A man who hates spaghetti code and having a big sharp axe is hunting me, so I had to clean up code

+++ lab2/1256A.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int min(int a, int b) {
+long long min(long long a, long long b) {
     return a < b ? a : b;
 }
